+++ syncmoney-papi-expansion/src/main/java/noietime/syncmoney/expansion/SyncmoneyExpansion.java
@@ -56,7 +56,7 @@ public class SyncmoneyExpansion extends PlaceholderExpansion {
 
     @Override
     public @NotNull String getVersion() {
-        return "1.1.1";
+        return "1.1.2";
     }
 
     @Override